requirements.txt
asgiref==3.6.0
boto3==1.26.47
botocore==1.29.47
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.1.14
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
ffmpeg==1.4
idna==3.4
jmespath==1.0.1
pdfminer.six==20221105
pdfplumber==0.7.6
Pillow==9.4.0
playsound==1.3.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
PyPDF3==1.0.6
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.7
python-vlc==3.0.18121
pyttsx3==2.90
pytz==2022.7
pywin32==305
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SpeechRecognition==3.9.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
Wand==0.6.10
